@@ -35,13 +35,13 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<revision>4.7.0-SNAPSHOT</revision>
-		<c8y.version>1020.127.0</c8y.version>
+		<c8y.version>1020.150.1</c8y.version>
 		<lombok.version>1.18.34</lombok.version>
 		<slf4j.version>1.7.36</slf4j.version>
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.target>21</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<spring-boot-dependencies.version>2.7.17</spring-boot-dependencies.version>
+		<spring-boot-dependencies.version>3.0.13</spring-boot-dependencies.version>
 	</properties>
 	<repositories>
 		<repository>
@@ -153,20 +153,20 @@
 				<artifactId>spring-aspects</artifactId>
 				<version>6.1.12</version>
 			</dependency>
-			<dependency>
+			<!-- <dependency>
 				<groupId>io.opentelemetry</groupId>
 				<artifactId>opentelemetry-bom</artifactId>
-				<version>1.40.0</version>
+				<version>1.44.1</version>
 				<type>pom</type>
 				<scope>import</scope>
-			</dependency>
-			<dependency>
+			</dependency> -->
+			<!-- <dependency>
 				<groupId>io.opentelemetry.instrumentation</groupId>
 				<artifactId>opentelemetry-instrumentation-bom</artifactId>
-				<version>2.6.0</version>
+				<version>2.8.0</version>
 				<type>pom</type>
 				<scope>import</scope>
-			</dependency>
+			</dependency> -->
 			<dependency>
 				<groupId>com.github.loki4j</groupId>
 				<artifactId>loki-logback-appender</artifactId>
