@@ -30,7 +30,7 @@ import lombok.Getter;
 import lombok.ToString;
 import dynamic.mapping.processor.model.RepairStrategy;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.io.Serializable;
 import java.util.List;
 import java.util.Map;
